Создаёт board-controller в который переносит логику из main
